Separate unreliable car into two files, fix drive method to override instead of add new method
