+++ prac_08/unreliable_car_test.py
@@ -1,22 +1,17 @@
-import random
-from prac_08.car import Car
+from prac_08.unreliable_car import UnreliableCar
+
+DISTANCE = 25
 
 
-class UnreliableCar(Car):
-    """A Car class that has a random chance of failing."""
-    reliability_chance = float(random.randint(0, 100))
-
-    def __init__(self, name, fuel, reliability=reliability_chance):
-        super().__init__(name, fuel)
-        self.reliability = reliability
-
-    def attempt_drive(self, distance):
-        if random.randint(0, 100) < self.reliability:
-            distance_driven = self.drive(distance)
-            print(f"Distance: {distance_driven}km")
-        else:
-            print("Distance: 0km")
+def main():
+    """A Car, but less reliable."""
+    bad_car = UnreliableCar("Crash-Test dummy", 100)
+    print(f"Driving {DISTANCE}km...")
+    bad_car.drive(DISTANCE)
+    if bad_car.odometer > 0:
+        print(f"Distance travelled: {bad_car.odometer}km")
+        print(f"Current stats: {bad_car.__str__()}")
 
 
-bad_car = UnreliableCar("Bad car", 100)
-bad_car.attempt_drive(50)
+main()
+
